requirements.txt
kafka-python
Flask~=1.1.2
peewee~=3.14.0
pymysql